+++ vpype_cli/metadata.py
@@ -7,7 +7,7 @@ import vpype as vp
 
 from .cli import cli
 from .decorators import global_processor, layer_processor
-from .types import LayerType, LengthType, multiple_to_layer_ids
+from .types import LayerType, LengthType, TextType, multiple_to_layer_ids
 
 __all__ = (
     "propset",
@@ -48,8 +48,8 @@ def _check_scope(
 
 
 @cli.command(group="Metadata")
-@click.argument("prop")
-@click.argument("value")
+@click.argument("prop", type=TextType())
+@click.argument("value", type=TextType())
 @click.option("--global", "-g", "global_flag", is_flag=True, help="Global mode.")
 @click.option("-l", "--layer", type=LayerType(accept_multiple=True), help="Target layer(s).")
 @click.option(
@@ -164,7 +164,7 @@ def proplist(document: vp.Document, global_flag: bool, layer: Optional[Union[int
 
 
 @cli.command(group="Metadata")
-@click.argument("prop")
+@click.argument("prop", type=TextType())
 @click.option("--global", "-g", "global_flag", is_flag=True, help="Global mode.")
 @click.option("-l", "--layer", type=LayerType(accept_multiple=True), help="Target layer(s).")
 @global_processor
@@ -198,7 +198,7 @@ def propget(
 
 
 @cli.command(group="Metadata")
-@click.argument("prop")
+@click.argument("prop", type=TextType())
 @click.option("--global", "-g", "global_flag", is_flag=True, help="Global mode.")
 @click.option("-l", "--layer", type=LayerType(accept_multiple=True), help="Target layer(s).")
 @global_processor
@@ -287,7 +287,7 @@ def penwidth(layer: vp.LineCollection, pen_width: float) -> vp.LineCollection:
 
 # noinspection PyShadowingNames
 @cli.command(group="Metadata")
-@click.argument("color", type=str)
+@click.argument("color", type=TextType())
 @layer_processor
 def color(layer: vp.LineCollection, color: str) -> vp.LineCollection:
     """Set the color for one or more layers.
@@ -315,7 +315,7 @@ def color(layer: vp.LineCollection, color: str) -> vp.LineCollection:
 
 # noinspection PyShadowingNames
 @cli.command(group="Metadata")
-@click.argument("name", type=str)
+@click.argument("name", type=TextType())
 @layer_processor
 def name(layer: vp.LineCollection, name: str) -> vp.LineCollection:
     """Set the name for one or more layers.
@@ -360,7 +360,7 @@ command. Check the documentation for more information on creating custom pen con
 
 
 @cli.command(group="Metadata", help=PENS_HELP_STRING)
-@click.argument("pen_config", metavar="CONF")
+@click.argument("pen_config", metavar="CONF", type=TextType())
 @global_processor
 def pens(document: vp.Document, pen_config: str) -> vp.Document: